microservices/ai-extraction-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.3.6
httpx==0.25.2
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
prometheus-client==0.19.0
structlog==23.2.0